Rename project to "fontobene-qt"
