+++ CHANGELOG.md
@@ -10,14 +10,3 @@ Possible log types:
 - `[removed]` for deprecated features removed in this release.
 - `[fixed]` for any bug fixes.
 - `[security]` to invite users to upgrade in case of vulnerabilities.
-
-
-### 0.2.0 (2021-01-01)
-
-- [added] Support for CMake (#10)
-- [fixed] Fixed wrong include path in example pkg-config file (#13)
-- [changed] Don't pass "SkipEmptyParts" when splitting polylines (#6)
-
-### 0.1.0 (2020-05-18)
-
-First release
