+++ README.md
@@ -1,19 +1,21 @@
-# C++/Qt5 FontoBene Parser
+# C++/Qt FontoBene Parser
 
-A header-only library to parse FontoBene stroke fonts with C++11/Qt5.
+A header-only library to parse FontoBene stroke fonts with C++11/Qt.
+
+Successor of `fontobene-qt5`, now supporting both Qt5 and Qt6.
 
 ## Installing
 
     $ mkdir build && cd build
     $ cmake .. -DCMAKE_INSTALL_PREFIX=/usr
     $ make -j8
-    $ tests/fontobene-qt5-tests
+    $ tests/fontobene-qt-tests
     $ make install
 
 ## pkg-config
 
-If you're packaging fontobene-qt5 and want an example pkg-config file, check
-out `fontobene-qt5.pc.example`. Depending on your distro, the `includedir`
+If you're packaging fontobene-qt and want an example pkg-config file, check
+out `fontobene-qt-5.pc.example`. Depending on your distro, the `includedir`
 might need adjusting.
 
 ## License
