+++ fontobene-qt-5.pc.example
@@ -1,9 +1,9 @@
 prefix=/usr
 includedir=${prefix}/include
 
-Name: fontobene-qt5
+Name: fontobene-qt
 Description: FontoBene parser for Qt5 (header-only)
-Version: 0.2.0
+Version: 1.0.0
 Libs:
 Cflags: -I${includedir}
 Requires: Qt5Core
