+++ fontobene-qt/glyph.h
