+++ fontobene-qt/glyphlist.h
