+++ fontobene-qt/glyphlistaccessor.h
