+++ fontobene-qt/glyphlistcache.h
