+++ fontobene-qt/helpers.h
