+++ tests/CMakeLists.txt
@@ -1,5 +1,5 @@
 # Test binary
-add_executable(fontobene_qt5_tests
+add_executable(fontobene_qt_tests
     main.cpp
 )
 
@@ -7,10 +7,10 @@ add_executable(fontobene_qt5_tests
 add_definitions(-DTESTFILEPATH="${CMAKE_CURRENT_SOURCE_DIR}/fonto.bene")
 
 # Fail on warnings
-target_compile_options(fontobene_qt5_tests PRIVATE -Werror -Wall -Wextra)
+target_compile_options(fontobene_qt_tests PRIVATE -Werror -Wall -Wextra)
 
 # Linking
-target_link_libraries(fontobene_qt5_tests fontobene_qt5 Qt5::Core)
+target_link_libraries(fontobene_qt_tests fontobene_qt Qt5::Core)
 
 # Output binary name
-set_target_properties(fontobene_qt5_tests PROPERTIES OUTPUT_NAME fontobene-qt5-tests)
+set_target_properties(fontobene_qt_tests PROPERTIES OUTPUT_NAME fontobene-qt-tests)
